Refactor NotePreviewList component and add resetScroll functionality
This commit refactors the NotePreviewList component to accept an optional onSelect prop. It also adds the resetScroll functionality to the App component, which is triggered when a note is selected in the NotePreviewList. These changes improve the user experience by resetting the scroll position when a new note is selected.